implementation of seperation widget of more_view
