+++ lib/features/more/more_view.dart
@@ -1,23 +1,21 @@
-// Flutter imports:
-import 'package:flutter/gestures.dart';
+// MoreView.dart
+
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
-
-// Package imports:
-import 'package:feature_discovery/feature_discovery.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:notredame/features/more/widget/contributors_list_tile.dart';
+import 'package:notredame/features/more/widget/help_list_tile.dart';
+import 'package:notredame/features/more/widget/in_app_review_list_tile';
+import 'package:notredame/features/more/widget/logout_list_tile.dart';
+import 'package:notredame/features/more/widget/open_source_licenses_list_tile.dart';
+import 'package:notredame/features/more/widget/privacy_policy_list_tile.dart';
+import 'package:notredame/features/more/widget/report_bug_list_title.dart';
+import 'package:notredame/features/more/widget/settings_list_tile.dart';
 import 'package:stacked/stacked.dart';
 
-// Project imports:
-import 'package:notredame/features/welcome/discovery/models/discovery_ids.dart';
-import 'package:notredame/features/app/navigation/router_paths.dart';
-import 'package:notredame/features/app/analytics/analytics_service.dart';
-import 'package:notredame/utils/utils.dart';
 import 'package:notredame/features/more/more_viewmodel.dart';
-import 'package:notredame/utils/locator.dart';
-import 'package:notredame/utils/app_theme.dart';
-import 'package:notredame/features/welcome/discovery/discovery_components.dart';
 import 'package:notredame/features/app/widgets/base_scaffold.dart';
+
 
 class MoreView extends StatefulWidget {
   @override
@@ -25,8 +23,6 @@ class MoreView extends StatefulWidget {
 }
 
 class _MoreViewState extends State<MoreView> {
-  final AnalyticsService _analyticsService = locator<AnalyticsService>();
-  static const String tag = "MoreView";
   bool isDiscoveryOverlayActive = false;
 
   @override
@@ -46,30 +42,6 @@ class _MoreViewState extends State<MoreView> {
         : Icon(icon);
   }
 
-  /// License text box
-  List<Widget> aboutBoxChildren(BuildContext context) {
-    final textStyle = Theme.of(context).textTheme.bodyMedium!;
-    return <Widget>[
-      const SizedBox(height: 24),
-      RichText(
-        text: TextSpan(
-          children: <TextSpan>[
-            TextSpan(
-                style: textStyle, text: AppIntl.of(context)!.flutter_license),
-            TextSpan(
-                style: textStyle.copyWith(color: Colors.blue),
-                text: AppIntl.of(context)!.flutter_website,
-                recognizer: TapGestureRecognizer()
-                  ..onTap = () => Utils.launchURL(
-                      AppIntl.of(context)!.flutter_website,
-                      AppIntl.of(context)!)),
-            TextSpan(style: textStyle, text: '.'),
-          ],
-        ),
-      ),
-    ];
-  }
-
   @override
   Widget build(BuildContext context) {
     return ViewModelBuilder<MoreViewModel>.reactive(
@@ -83,175 +55,18 @@ class _MoreViewState extends State<MoreView> {
             body: ListView(
               padding: EdgeInsets.zero,
               children: [
-                ListTile(
-                    title: Text(AppIntl.of(context)!.more_about_applets_title),
-                    leading: _buildDiscoveryFeatureDescriptionWidget(
-                        context,
-                        Hero(
-                          tag: 'about',
-                          child: Image.asset(
-                            "assets/images/favicon_applets.png",
-                            height: 24,
-                            width: 24,
-                          ),
-                        ),
-                        DiscoveryIds.detailsMoreThankYou,
-                        model),
-                    onTap: () {
-                      _analyticsService.logEvent(tag, "About App|ETS clicked");
-                      model.navigationService.pushNamed(RouterPaths.about);
-                    }),
-                ListTile(
-                    title: Text(AppIntl.of(context)!.more_report_bug),
-                    leading: _buildDiscoveryFeatureDescriptionWidget(
-                        context,
-                        getProperIconAccordingToTheme(Icons.bug_report),
-                        DiscoveryIds.detailsMoreBugReport,
-                        model),
-                    onTap: () {
-                      _analyticsService.logEvent(tag, "Report a bug clicked");
-                      model.navigationService.pushNamed(RouterPaths.feedback);
-                    }),
-                ListTile(
-                    title: Text(AppIntl.of(context)!.in_app_review_title),
-                    leading: const Icon(Icons.rate_review),
-                    onTap: () {
-                      _analyticsService.logEvent(tag, "Rate us clicked");
-                      MoreViewModel.launchInAppReview();
-                    }),
-                ListTile(
-                    title: Text(AppIntl.of(context)!.more_contributors),
-                    leading: _buildDiscoveryFeatureDescriptionWidget(
-                        context,
-                        getProperIconAccordingToTheme(Icons.people_outline),
-                        DiscoveryIds.detailsMoreContributors,
-                        model),
-                    onTap: () {
-                      _analyticsService.logEvent(tag, "Contributors clicked");
-                      model.navigationService
-                          .pushNamed(RouterPaths.contributors);
-                    }),
-                ListTile(
-                    title: Text(AppIntl.of(context)!.more_open_source_licenses),
-                    leading: const Icon(Icons.code),
-                    onTap: () {
-                      _analyticsService.logEvent(tag, "Rate us clicked");
-                      Navigator.of(context).push(PageRouteBuilder(
-                        pageBuilder: (context, _, __) => AboutDialog(
-                          applicationIcon: Padding(
-                            padding: const EdgeInsets.all(8.0),
-                            child: SizedBox(
-                                width: 75,
-                                child: Image.asset(
-                                    'assets/images/favicon_applets.png')),
-                          ),
-                          applicationName:
-                              AppIntl.of(context)!.more_open_source_licenses,
-                          applicationVersion: model.appVersion,
-                          applicationLegalese:
-                              '\u{a9} ${DateTime.now().year} App|ETS',
-                          children: aboutBoxChildren(context),
-                        ),
-                        opaque: false,
-                      ));
-                    }),
-                if (model.privacyPolicyToggle)
-                  ListTile(
-                      title: Text(AppIntl.of(context)!.privacy_policy),
-                      leading: const Icon(Icons.privacy_tip),
-                      onTap: () {
-                        _analyticsService.logEvent(
-                            tag, "Confidentiality clicked");
-                        MoreViewModel.launchPrivacyPolicy();
-                      }),
-                ListTile(
-                    title: Text(AppIntl.of(context)!.need_help),
-                    leading: _buildDiscoveryFeatureDescriptionWidget(
-                        context,
-                        getProperIconAccordingToTheme(Icons.question_answer),
-                        DiscoveryIds.detailsMoreFaq,
-                        model),
-                    onTap: () {
-                      _analyticsService.logEvent(tag, "FAQ clicked");
-                      model.navigationService.pushNamed(RouterPaths.faq,
-                          arguments: Utils.getColorByBrightness(
-                              context, Colors.white, AppTheme.primaryDark));
-                    }),
-                ListTile(
-                    title: Text(AppIntl.of(context)!.settings_title),
-                    leading: _buildDiscoveryFeatureDescriptionWidget(
-                        context,
-                        getProperIconAccordingToTheme(Icons.settings),
-                        DiscoveryIds.detailsMoreSettings,
-                        model),
-                    onTap: () {
-                      _analyticsService.logEvent(tag, "Settings clicked");
-                      model.navigationService.pushNamed(RouterPaths.settings);
-                    }),
-                ListTile(
-                  title: Text(AppIntl.of(context)!.more_log_out),
-                  leading: const Icon(Icons.logout),
-                  onTap: () => Navigator.of(context).push(
-                    PageRouteBuilder(
-                      pageBuilder: (context, _, __) => AlertDialog(
-                        title: Text(
-                          AppIntl.of(context)!.more_log_out,
-                          style: const TextStyle(color: Colors.red),
-                        ),
-                        content: Text(AppIntl.of(context)!
-                            .more_prompt_log_out_confirmation),
-                        actions: [
-                          TextButton(
-                              onPressed: () async {
-                                _analyticsService.logEvent(
-                                    tag, "Log out clicked");
-                                model.logout();
-                              },
-                              child: Text(AppIntl.of(context)!.yes)),
-                          TextButton(
-                              onPressed: () => Navigator.of(context).pop(),
-                              child: Text(AppIntl.of(context)!.no))
-                        ],
-                      ),
-                      opaque: false,
-                    ),
-                  ),
-                ),
+                const AboutListTile(),
+                ReportBugListTile(),
+                InAppReviewListTile(),
+                ContributorsListTile(),
+                OpenSourceLicensesListTile(),
+                if (model.privacyPolicyToggle) PrivacyPolicyListTile(),
+                HelpListTile(),
+                SettingsListTile(),
+                LogoutListTile(),
               ],
             ),
           );
-        });
-  }
-
-  DescribedFeatureOverlay _buildDiscoveryFeatureDescriptionWidget(
-      BuildContext context,
-      Widget icon,
-      String featuredId,
-      MoreViewModel model) {
-    final discovery = getDiscoveryByFeatureId(
-        context, DiscoveryGroupIds.pageMore, featuredId);
-
-    return DescribedFeatureOverlay(
-        overflowMode: OverflowMode.wrapBackground,
-        contentLocation: ContentLocation.below,
-        featureId: discovery.featureId,
-        title: Text(discovery.title, textAlign: TextAlign.justify),
-        description: discovery.details,
-        backgroundColor: AppTheme.appletsDarkPurple,
-        tapTarget: icon,
-        pulseDuration: const Duration(seconds: 5),
-        child: icon,
-        onComplete: () {
-          setState(() {
-            isDiscoveryOverlayActive = false;
-          });
-          return model.discoveryCompleted();
-        },
-        onOpen: () async {
-          setState(() {
-            isDiscoveryOverlayActive = true;
-          });
-          return true;
         });
   }
 }
